Make definition inputs fully controlled
Previously, the textarea was uncontrolled (state was managed internally
and propagated up with an onchange callback) which meant there was a
duplication in state since <App> maintained the deserialized version and
the textarea maintained the serialized version.

Now the textarea is fully controlled and the state lives only in <App>
in the serialized format, which is only deserialized when passing it
down to the <RecipeInput>.

This makes the value of the {Ingredient,Unit}DefinitionsInput components
pretty low so I decided to just inline them in RecipeDefinitionsInput.